implement Liquibase OpenSearch Spring Boot Starter
this is currently based on `spring-data-opensearch` v1.8.1 and not
v2.0.0 which was released a few days ago because the current v0.0.1
release of `liquibase-opensearch` is based on `opensearch-java` v2 and
not the new v3 which is being used by the new `spring-data-opensearch`
release.

thus the goal is to have a v0.0.1 release also of the new SB starter
basd on `opensearch-java` v2 and then a v0.1.0 release of both based on
the new major release of the opensearch client library.

resolves liquibase/liquibase-opensearch#86