update to `spring-data-opensearch` v2.0.0
due to the use of `@ServiceConnection` in the tests (which we could
remove) we also need the newer `opensearch-testcontainers` version (as
that is used in the newer SB integration) which in turn forces us to
build only with JDK 21. technically this is only for the test sources,
however the way liquibase workflows work we have to use the same version
for main and test sources and thus drop support for older JDKs.

this is the first notable change since the initial v0.0.1 release,
accordingly i've now introduced a changelog.