support custom index names for changelog & lock indices
when using the CLI several options exist to set these settings:
* https://docs.liquibase.com/parameters/database-changelog-table-name.html
* https://docs.liquibase.com/parameters/database-changelog-lock-table-name.html

these are however not used with Spring Boot. for "normal" Liquibase the
`databaseChangeLogTable` and `databaseChangeLogLockTable` parameters
under `spring.liquibase` can be used.

to allow the co-existence of OpenSearch and non-OpenSearch SB Liquibase
integrations we have not re-used these parameters and instead added new
SB parameters under `opensearch.liquibase`. as OpenSearch contains
indices rather than tables the parameter names have been slightly
adjusted to reflect this. the default value is the same as for the
non-OpenSearch case, just that the default is already lowercase as it'll
be transformed to this anyway by `liquibase-opensearch` since OpenSearch
only supports lowercase index names.

resolves #44